Split sample into avoid/do dirs to standardize
