+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-avoid/groovy/library/build.gradle
@@ -1,0 +1,41 @@
+plugins {
+    id 'java-library'
+}
+
+// tag::bad-classpath-printer[]
+dependencies {
+    runtimeOnly(project(":library")) // <1>
+}
+
+abstract class BadClasspathPrinter extends DefaultTask {
+    @InputFiles
+    Set<File> classpath = [] as Set // <2>
+
+    protected int calculateDigest(File fileOrDirectory) {
+        if (!fileOrDirectory.exists()) {
+            throw new IllegalArgumentException("File or directory $fileOrDirectory doesn't exist")
+        }
+        return 0 // actual implementation is stripped
+    }
+
+    @TaskAction
+    void run() {
+        logger.lifecycle(
+            classpath.collect { file ->
+                def digest = calculateDigest(file) // <4>
+                "$file#$digest"
+            }.join("\n")
+        )
+    }
+}
+
+tasks.register("badClasspathPrinter", BadClasspathPrinter) {
+    classpath = configurations.named("runtimeClasspath").get().resolve() // <3>
+}
+// end::bad-classpath-printer[]
+
+tasks.named("jar").configure {
+    doLast {
+        logger.lifecycle("jar task was executed")
+    }
+}
