+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-avoid/groovy/library/src/main/java/Sample.java
