+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-avoid/groovy/settings.gradle
