+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-avoid/kotlin/library/src/main/java/Sample.java
