+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-avoid/kotlin/settings.gradle.kts
