+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-avoid/tests/avoidResolvingConfigurationsManually-avoid.sample.conf
@@ -1,0 +1,6 @@
+executable: gradle
+args: ":library:badClasspathPrinter"
+expected-output-file: avoidResolvingConfigurationsManually-avoid.out
+expect-failure: true
+allow-additional-output: true
+allow-disordered-output: true
