+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-do/groovy/settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = "avoid-resolving-configurations-manually"
+
+include(":library")
