+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-do/kotlin/library/src/main/java/Sample.java
@@ -1,0 +1,3 @@
+package sample;
+
+class Sample {}
