+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-do/kotlin/settings.gradle.kts
@@ -1,0 +1,3 @@
+rootProject.name = "avoid-resolving-configurations-manually"
+
+include(":library")
