+++ platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually-do/tests/avoidResolvingConfigurationsManually-do.sample.conf
@@ -1,0 +1,3 @@
+executable: gradle
+args: ":library:goodClasspathPrinter"
+expected-output-file: avoidResolvingConfigurationsManually-do.out
