--- platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually/groovy/library/build.gradle
@@ -1,9 +1,0 @@
-plugins {
-    id 'java-library'
-}
-
-tasks.jar.configure {
-    doLast {
-        logger.lifecycle("jar task was executed")
-    }
-}
