--- platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually/kotlin/library/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-    `java-library`
-}
-
-tasks.jar.configure {
-    doLast {
-        logger.lifecycle("jar task was executed")
-    }
-}
