--- platforms/documentation/docs/src/snippets/bestPractices/avoidResolvingConfigurationsManually/tests/avoidResolvingConfigurationsManually.sample.conf
@@ -1,12 +1,0 @@
-commands: [{
-    executable: gradle
-    args: badClasspathPrinter
-    expected-output-file: avoidResolvingConfigurationsManually.bad.out
-    expect-failure: true
-    allow-additional-output: true
-    allow-disordered-output: true
-}, {
-    executable: gradle
-    args: goodClasspathPrinter
-    expected-output-file: avoidResolvingConfigurationsManually.good.out
-}]
